requirements.txt
streamlit==1.42.0
streamlit-extras==0.5.0
requests==2.31.0
PyPDF2==3.0.1
python-dotenv==1.0.1